refactor: EndpointRefresher 분리
